requirements.txt
Django==1.6.5
argparse==1.2.1
djangorestframework==2.3.14
requests==2.3.0
wsgiref==0.1.2
